backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
